Edited ex_net.py and test_loadW.py
Worked with Marina to update ex_net.py and test_loadW.py to read in W
and stats pickle files.

test_loadW.py is good for understanding how to read pickle files and how
to edit the stats dictionary, which will be useful after running
simulation and recording synchronicity results.